+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/external/samples/readme.md
@@ -3,7 +3,7 @@
 No Team-specific code should be placed or modified in this ``.../samples`` folder.
 
 Samples should be Copied from here, and then Pasted into the team's 
-[/TeamCode/src/main/java/org/firstinspires/ftc/teamcode](../../../../../../../../../../TeamCode/src/main/java/org/firstinspires/ftc/Team19567)
+[/TeamCode/src/main/java/org/firstinspires/ftc/teamcode](../../../../../../../../../../TeamCode/src/main/java/org/firstinspires/ftc/teamcode)
  folder, using the Android Studio cut and paste commands.  This automatically changes all file and
 class names to be consistent.  From there, the sample can be modified to suit the team's needs.
