+++ TeamCode/build.gradle
@@ -16,7 +16,7 @@ apply from: '../build.common.gradle'
 apply from: '../build.dependencies.gradle'
 
 android {
-    namespace = 'org.firstinspires.ftc.Team19567'
+    namespace = 'org.firstinspires.ftc.teamcode'
     packagingOptions {
         jniLibs.useLegacyPackaging true
     }
@@ -28,7 +28,7 @@ dependencies {
 
     implementation 'org.apache.commons:commons-math3:3.6.1'
 //    implementation 'com.acmerobotics.roadrunner:core:0.5.4'
-    implementation 'com.acmerobotics.roadrunner:core:1.0.0-beta0'
+    implementation 'com.acmerobotics.roadrunner:core:0.5.6'
 //    implementation 'org.openftc:easyopencv:1.5.1'
     implementation 'org.openftc:opencv-repackaged:4.5.3-B'
 }
