+++ build.dependencies.gradle
@@ -18,10 +18,12 @@ dependencies {
     implementation 'org.firstinspires.ftc:Hardware:8.1.0'
     implementation 'org.firstinspires.ftc:FtcCommon:8.1.0'
     implementation 'org.tensorflow:tensorflow-lite-task-vision:0.2.0'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
+//    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'org.firstinspires.ftc:gameAssets-FreightFrenzy:1.0.0'
 
-    implementation 'com.acmerobotics.dashboard:dashboard:0.4.3'
+//    implementation 'com.acmerobotics.dashboard:dashboard:0.4.3'
+    implementation 'com.acmerobotics.dashboard:dashboard:0.4.6'
     implementation 'org.openftc:easyopencv:1.5.1'
     implementation 'org.firstinspires.ftc:gameAssets-PowerPlay:1.0.0'
 }
