+++ settings.gradle
@@ -1,4 +1,3 @@
 include ':FtcRobotController'
 include ':TeamCode'
 include ':MeepMeepTesting'
-include ':SimpleRobotCode'
